style(realizer): make the realizer more concise
This was actually a part of a larger investigation. When I was double checking the realizer, I somehow got a different number for the length of the results, down to 1,113,316 from 1,122,348. But when I went to investigate again, it was back to 1,122,348, even the previous commits. It's very strange, but I am pretty confident that the code is correct, and so I've retroactively pretended that this never happened, but this commit will tell you of this spooky fright.